refactor testers
